refactor(routes): extract /feedback into its own module (god-file decomposition)
Feedback routes moved verbatim from mycelium.js into server/routes/feedback.js
(handler bodies byte-identical, async preserved), wired via registerFeedbackRoutes.
Route contract unchanged (route-manifest: 284 routes), vitest green. Lucy (GLM-5.2),
byte-reviewed by m5Max.
